requirements.txt
annotated-types==0.7.0
anyio==4.10.0
certifi==2025.8.3
click==8.2.1
faiss-cpu==1.8.0.post1
fastapi==0.115.0
h11==0.16.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.27.2
idna==3.10
numpy==1.26.4
ollama==0.3.3
orjson==3.10.7
packaging==25.0
pydantic==2.9.2
pydantic_core==2.23.4
google-generativeai==0.7.2
pypdf==5.0.1
python-dotenv==1.0.1
PyYAML==6.0.2
sniffio==1.3.1
starlette==0.38.6
typing_extensions==4.15.0
uvicorn==0.30.6
uvloop==0.21.0
watchfiles==1.1.0
websockets==15.0.1
